feat: handle missing params error

+++ src/main/java/carijodoh/service/ChatService.java
@@ -20,6 +20,11 @@ public class ChatService {
             @WebParam(name = "userIdReceiver") int userIdReceiver,
             @WebParam(name = "message") String message,
             @WebParam(name = "apiKey") String apiKey){
+        // Handle parameters
+        if(notValidInput(message) || notValidInput(apiKey) || userIdReceiver == 0 || userIdSender == 0){
+            return "Missing params";
+        }
+
         if (!apiKey.equals(Dotenv.load().get("API_KEY_PHP")) && !apiKey.equals(Dotenv.load().get("API_KEY_REST"))) {
             return "Not authorized";
         } else {
@@ -32,6 +37,11 @@ public class ChatService {
             @WebParam(name = "userIdSender") int userIdSender,
             @WebParam(name = "userIdReceiver") int userIdReceiver,
             @WebParam(name = "apiKey") String apiKey){
+        // Handle parameters
+        if(notValidInput(apiKey) || userIdReceiver == 0 || userIdSender == 0){
+            return new DataChat();
+        }
+
         if (!apiKey.equals(Dotenv.load().get("API_KEY_PHP")) && !apiKey.equals(Dotenv.load().get("API_KEY_REST"))) {
             return new DataChat();
         } else {
@@ -44,10 +54,19 @@ public class ChatService {
             @WebParam(name = "userIdSender") int userIdSender,
             @WebParam(name = "userIdReceiver") int userIdReceiver,
             @WebParam(name = "apiKey") String apiKey){
+        // Handle parameters
+        if(notValidInput(apiKey) || userIdReceiver == 0 || userIdSender == 0){
+            return "Missing params";
+        }
+
         if (!apiKey.equals(Dotenv.load().get("API_KEY_PHP")) && !apiKey.equals(Dotenv.load().get("API_KEY_REST"))) {
             return "Not authorized";
         } else {
             return chatRepository.deleteChat(userIdSender, userIdReceiver);
         }
+    }
+
+    private boolean notValidInput(String input) {
+        return input == null || input.trim().isEmpty();
     }
 }
